测试成功

+++ cq_windows.go
@@ -18,7 +18,7 @@ func _on_enable() int32 { return Enable() }
 //AddLog 增加运行日志
 func AddLog(priority int32, logtype, reason string) int32 {
 	return int32(C.CQ_addLog(
-		priority,
+		C.int32_t(priority),
 		C.CString(logtype),
 		C.CString(reason),
 	))
@@ -27,6 +27,6 @@ func AddLog(priority int32, logtype, reason string) int32 {
 //SendPrivateMsg 发送私聊消息
 func SendPrivateMsg(qq int64, msg string) int32 {
 	return int32(C.CQ_sendPrivateMsg(
-		qq, C.CString(msg),
+		C.int64_t(qq), C.CString(msg),
 	))
 }
